Target generation scripts updated.
The spot stimulus required a fix to the number of pixels in the inner
and out region. The fix in this commit reflects a change that uses
different amplitudes for the inner and out regions.

The haar2D stimulus is generalized to produce the appropriate stimulus
for a dc = 0 and dc = N.