A change to the way similarity is calculated in the amplitude domain.
The previous version of the similarity cropped the Fourier image to
remove spatial frequencies that we presume are not present in our
target. This was used primarily to counter any spatial frequencies
included due only to the padding. This method is ineffective. Instead,
our measure now uses the entire Fourier image in the measure of phase
independent similarity.