Phase invariant similarity change.
We previously padded to the nearest power of two. Then it was observed
that pad size changes the similarity value, especially for broadband
targets such as the occluding edge targets. Thus, we now pad to 128 for
all targets. Furthermore, similarity is now taken in the frequency
domain over the entire Fourier image.